pyproject.toml
[project]
name = "database-testing"
version = "0.1.0"
description = "Testing the performance of different databases"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clickhouse-connect>=0.8.16",
    "clickhouse-driver>=0.2.9",
    "crate>=2.0.0",
    "dask>=2023.4.0",
    "influxdb3-python>=0.12.0",
    "numpy<2.0.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.1.0",
    "questdb>=2.0.3",
    "requests>=2.32.3",
    "sqlalchemy-cratedb>=0.41.0",
    "taos>=1.0.1",
    "taospy>=2.8.0",
]

[dependency-groups]
dev = [
    "pandas-stubs>=2.2.3.250308",
    "plotly>=6.0.1",
    "types-psycopg2>=2.9.21.20250318",
    "types-requests>=2.32.0.20250328",
]
